add texture.

+++ path-tracing-core/emitters/point_emitter.cpp
@@ -6,7 +6,7 @@ path_tracing::core::emitters::point_emitter::point_emitter(const vector3& intens
 }
 
 path_tracing::core::emitters::emitter_gpu_buffer path_tracing::core::emitters::point_emitter::gpu_buffer(
-	const transform& transform, size_t index) const noexcept
+	const transform& transform, uint32 index) const noexcept
 {
 	emitter_gpu_buffer buffer;
 
@@ -14,7 +14,7 @@ path_tracing::core::emitters::emitter_gpu_buffer path_tracing::core::emitters::p
 	buffer.position = transform_point(transform, vector3(0));
 	buffer.type = type();
 	buffer.delta = 1;
-	buffer.index = static_cast<uint32>(index);
+	buffer.index = index;
 	
 	return buffer;
 }
